refactoring idle workload
