+++ vmm_tests/vmm_tests/tests/tests/multiarch.rs
@@ -23,7 +23,6 @@ use petri::pipette::cmd;
 use petri_artifacts_common::tags::MachineArch;
 use petri_artifacts_common::tags::OsFlavor;
 use petri_artifacts_vmm_test::artifacts::test_vmgs::VMGS_WITH_BOOT_ENTRY;
-use serde_json::to_string_pretty;
 use std::str::FromStr;
 use std::time::Duration;
 use vmm_test_macros::openvmm_test;
@@ -839,28 +838,7 @@ async fn validate_mnf_usage_in_guest(
 async fn memory_validation_tdx_2_proc_no_agent<T: PetriVmmBackend>(
     config: PetriVmBuilder<T>,
 ) -> anyhow::Result<()> {
-    let mut vm = config
-        .with_processor_topology(ProcessorTopology {
-            vp_count: 2,
-            ..Default::default()
-        })
-        .with_memory(MemoryConfig {
-            startup_bytes: 16 * (1024 * 1024 * 1024),
-            dynamic_memory_range: None,
-        })
-        .run_without_agent()
-        .await?;
-    std::thread::sleep(Duration::from_secs(60));
-    let vtl2_agent = vm.wait_for_vtl2_agent().await?;
-    let memstat = memstat::MemStat::new(&vtl2_agent).await;
-    tracing::info!(
-        "MEMSTAT_START:\n{}\n:MEMSTAT_END",
-        to_string_pretty(&memstat).unwrap()
-    );
-
-    vm.send_enlightened_shutdown(ShutdownKind::Shutdown).await?;
-    vm.wait_for_clean_teardown().await?;
-    assert!(memstat.compare_to_baseline("intel-tdx", "2vp"));
+    memstat::idle_test::<T>(config, "intel-tdx", 2, 16, 10).await?;
 
     Ok(())
 }
@@ -870,28 +848,7 @@ async fn memory_validation_tdx_2_proc_no_agent<T: PetriVmmBackend>(
 async fn memory_validation_tdx_64_proc_no_agent<T: PetriVmmBackend>(
     config: PetriVmBuilder<T>,
 ) -> anyhow::Result<()> {
-    let mut vm = config
-        .with_processor_topology(ProcessorTopology {
-            vp_count: 64,
-            ..Default::default()
-        })
-        .with_memory(MemoryConfig {
-            startup_bytes: 16 * (1024 * 1024 * 1024),
-            dynamic_memory_range: None,
-        })
-        .run_without_agent()
-        .await?;
-    std::thread::sleep(Duration::from_secs(60));
-    let vtl2_agent = vm.wait_for_vtl2_agent().await?;
-    let memstat = memstat::MemStat::new(&vtl2_agent).await;
-    tracing::info!(
-        "MEMSTAT_START:\n{}\n:MEMSTAT_END",
-        to_string_pretty(&memstat).unwrap()
-    );
-
-    vm.send_enlightened_shutdown(ShutdownKind::Shutdown).await?;
-    vm.wait_for_clean_teardown().await?;
-    assert!(memstat.compare_to_baseline("intel-tdx", "64vp"));
+    memstat::idle_test::<T>(config, "intel-tdx", 64, 16, 15).await?;
 
     Ok(())
 }
@@ -901,28 +858,7 @@ async fn memory_validation_tdx_64_proc_no_agent<T: PetriVmmBackend>(
 async fn memory_validation_snp_2_proc_no_agent<T: PetriVmmBackend>(
     config: PetriVmBuilder<T>,
 ) -> anyhow::Result<()> {
-    let mut vm = config
-        .with_processor_topology(ProcessorTopology {
-            vp_count: 2,
-            ..Default::default()
-        })
-        .with_memory(MemoryConfig {
-            startup_bytes: 16 * (1024 * 1024 * 1024),
-            dynamic_memory_range: None,
-        })
-        .run_without_agent()
-        .await?;
-    std::thread::sleep(Duration::from_secs(60));
-    let vtl2_agent = vm.wait_for_vtl2_agent().await?;
-    let memstat = memstat::MemStat::new(&vtl2_agent).await;
-    tracing::info!(
-        "MEMSTAT_START:\n{}\n:MEMSTAT_END",
-        to_string_pretty(&memstat).unwrap()
-    );
-
-    vm.send_enlightened_shutdown(ShutdownKind::Shutdown).await?;
-    vm.wait_for_clean_teardown().await?;
-    assert!(memstat.compare_to_baseline("amd-snp", "2vp"));
+    memstat::idle_test(config, "amd-snp", 2, 16, 10).await?;
 
     Ok(())
 }
@@ -932,28 +868,7 @@ async fn memory_validation_snp_2_proc_no_agent<T: PetriVmmBackend>(
 async fn memory_validation_snp_64_proc_no_agent<T: PetriVmmBackend>(
     config: PetriVmBuilder<T>,
 ) -> anyhow::Result<()> {
-    let mut vm = config
-        .with_processor_topology(ProcessorTopology {
-            vp_count: 64,
-            ..Default::default()
-        })
-        .with_memory(MemoryConfig {
-            startup_bytes: 16 * (1024 * 1024 * 1024),
-            dynamic_memory_range: None,
-        })
-        .run_without_agent()
-        .await?;
-    std::thread::sleep(Duration::from_secs(60));
-    let vtl2_agent = vm.wait_for_vtl2_agent().await?;
-    let memstat = memstat::MemStat::new(&vtl2_agent).await;
-    tracing::info!(
-        "MEMSTAT_START:\n{}\n:MEMSTAT_END",
-        to_string_pretty(&memstat).unwrap()
-    );
-
-    vm.send_enlightened_shutdown(ShutdownKind::Shutdown).await?;
-    vm.wait_for_clean_teardown().await?;
-    assert!(memstat.compare_to_baseline("amd-snp", "64vp"));
+    memstat::idle_test(config, "amd-snp", 64, 16, 15).await?;
 
     Ok(())
 }
@@ -963,31 +878,7 @@ async fn memory_validation_snp_64_proc_no_agent<T: PetriVmmBackend>(
 async fn memory_validation_x64_2_proc<T: PetriVmmBackend>(
     config: PetriVmBuilder<T>,
 ) -> anyhow::Result<()> {
-    let (mut vm, agent) = config
-        .with_processor_topology({
-            ProcessorTopology {
-                vp_count: 2,
-                ..Default::default()
-            }
-        })
-        .with_memory({
-            MemoryConfig {
-                startup_bytes: 16 * (1024 * 1024 * 1024),
-                dynamic_memory_range: None,
-            }
-        })
-        .run()
-        .await?;
-    std::thread::sleep(Duration::from_secs(60));
-    let vtl2_agent = vm.wait_for_vtl2_agent().await?;
-    let memstat = memstat::MemStat::new(&vtl2_agent).await;
-    tracing::info!(
-        "MEMSTAT_START:\n{}\n:MEMSTAT_END",
-        to_string_pretty(&memstat).unwrap()
-    );
-    agent.power_off().await?;
-    vm.wait_for_teardown().await?;
-    memstat.compare_to_baseline("intel-x64", "2vp");
+    memstat::idle_test(config, "intel-x64", 2, 16, 10).await?;
 
     Ok(())
 }
@@ -997,32 +888,7 @@ async fn memory_validation_x64_2_proc<T: PetriVmmBackend>(
 async fn memory_validation_x64_32_proc<T: PetriVmmBackend>(
     config: PetriVmBuilder<T>,
 ) -> anyhow::Result<()> {
-    let (mut vm, agent) = config
-        .with_processor_topology({
-            ProcessorTopology {
-                vp_count: 32,
-                ..Default::default()
-            }
-        })
-        .with_memory({
-            MemoryConfig {
-                startup_bytes: 16 * (1024 * 1024 * 1024),
-                dynamic_memory_range: None,
-            }
-        })
-        .run()
-        .await?;
-    std::thread::sleep(Duration::from_secs(60));
-    let vtl2_agent = vm.wait_for_vtl2_agent().await?;
-    let memstat = memstat::MemStat::new(&vtl2_agent).await;
-    tracing::info!(
-        "MEMSTAT_START:\n{}\n:MEMSTAT_END",
-        to_string_pretty(&memstat).unwrap()
-    );
-
-    agent.power_off().await?;
-    vm.wait_for_teardown().await?;
-    memstat.compare_to_baseline("intel-x64", "32vp");
+    memstat::idle_test(config, "intel-x64", 32, 16, 10).await?;
 
     Ok(())
 }
@@ -1032,32 +898,7 @@ async fn memory_validation_x64_32_proc<T: PetriVmmBackend>(
 async fn memory_validation_arm_2_proc<T: PetriVmmBackend>(
     config: PetriVmBuilder<T>,
 ) -> anyhow::Result<()> {
-    let (mut vm, agent) = config
-        .with_processor_topology({
-            ProcessorTopology {
-                vp_count: 2,
-                ..Default::default()
-            }
-        })
-        .with_memory({
-            MemoryConfig {
-                startup_bytes: 16 * (1024 * 1024 * 1024),
-                dynamic_memory_range: None,
-            }
-        })
-        .run()
-        .await?;
-    std::thread::sleep(Duration::from_secs(60));
-    let vtl2_agent = vm.wait_for_vtl2_agent().await?;
-    let memstat = memstat::MemStat::new(&vtl2_agent).await;
-    tracing::info!(
-        "MEMSTAT_START:\n{}\n:MEMSTAT_END",
-        to_string_pretty(&memstat).unwrap()
-    );
-    agent.power_off().await?;
-    vm.wait_for_teardown().await?;
-
-    assert!(memstat.compare_to_baseline("aarch64", "2vp"));
+    memstat::idle_test(config, "aarch64", 2, 16, 10).await?;
 
     Ok(())
 }
@@ -1067,32 +908,7 @@ async fn memory_validation_arm_2_proc<T: PetriVmmBackend>(
 async fn memory_validation_arm_64_proc<T: PetriVmmBackend>(
     config: PetriVmBuilder<T>,
 ) -> anyhow::Result<()> {
-    let (mut vm, agent) = config
-        .with_processor_topology({
-            ProcessorTopology {
-                vp_count: 64,
-                ..Default::default()
-            }
-        })
-        .with_memory({
-            MemoryConfig {
-                startup_bytes: 16 * (1024 * 1024 * 1024),
-                dynamic_memory_range: None,
-            }
-        })
-        .run()
-        .await?;
-    std::thread::sleep(Duration::from_secs(60));
-    let vtl2_agent = vm.wait_for_vtl2_agent().await?;
-    let memstat = memstat::MemStat::new(&vtl2_agent).await;
-    tracing::info!(
-        "MEMSTAT_START:\n{}\n:MEMSTAT_END",
-        to_string_pretty(&memstat).unwrap()
-    );
-
-    agent.power_off().await?;
-    vm.wait_for_teardown().await?;
-    assert!(memstat.compare_to_baseline("aarch64", "64vp"));
+    memstat::idle_test(config, "aarch64", 64, 16, 15).await?;
 
     Ok(())
 }
